# docs/swagger.yaml
basePath: /
definitions:
...
  auth.RequestRole:
    properties:
      description:
        example: Hak akses penuh untuk mengelola data master
        type: string
      name:
        example: Admin Desa
        type: string
    required:
    - name
    type: object
...
  master.RequestSumberDana:
    properties:
      namaSumber:
        example: Dana Desa (DD)
        type: string
    required:
    - namaSumber
    type: object
  response.Base:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
...
  transaction.RequestUsulanProyek:
    properties:
      lokasi:
        example: Dusun Sukamaju RT 01
        type: string
      namaProyek:
        example: Pembangunan Gorong-Gorong
        type: string
      nilaiRab:
        example: 45000000
        type: number
      satuan:
        example: Meter
        type: string
      statusSifat:
        example: Reguler
        type: string
      sumberDanaId:
        example: masukkan-uuid-sumber-dana-disini
        type: string
      tahunAnggaran:
        example: 2025
        type: integer
      volume:
        example: 150.5
        type: number
    required:
    - lokasi
    - namaProyek
    - nilaiRab
    - statusSifat
    - tahunAnggaran
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Backend untuk Sistem Pendukung Keputusan (SIPROPEDA) - Metode TOPSIS.
  title: SIPROPEDA API
  version: "1.0"
paths:
...
  /v1/roles:
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil semua data Role
      tags:
      - Roles
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role yang akan ditambahkan
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RequestRole'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Base'
      summary: Tambah data Role
      tags:
      - Roles
  /v1/roles/{id}:
    delete:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Role
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Hapus data Role (Soft Delete)
      tags:
      - Roles
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Role
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil detail Role by ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Role
        in: path
        name: id
        required: true
        type: string
      - description: Role yang akan diedit
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RequestRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Update data Role
      tags:
      - Roles
  /v1/roles/all:
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil semua data Role
      tags:
      - Roles
  /v1/sumber-dana:
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil semua data Sumber Dana
      tags:
      - Sumber Dana
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data Sumber Dana
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/master.RequestSumberDana'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Base'
      summary: Tambah data Sumber Dana
      tags:
      - Sumber Dana
  /v1/sumber-dana/{id}:
    delete:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Sumber Dana
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Hapus data Sumber Dana (Soft Delete)
      tags:
      - Sumber Dana
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Sumber Dana
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil detail Sumber Dana by ID
      tags:
      - Sumber Dana
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Sumber Dana
        in: path
        name: id
        required: true
        type: string
      - description: Data yang akan diedit
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/master.RequestSumberDana'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Update data Sumber Dana
      tags:
      - Sumber Dana
...
  /v1/usulan-proyek:
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil semua data Usulan Proyek
      tags:
      - Usulan Proyek
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data Usulan Proyek
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/transaction.RequestUsulanProyek'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Base'
      summary: Tambah data Usulan Proyek
      tags:
      - Usulan Proyek
  /v1/usulan-proyek/{id}:
    delete:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Usulan Proyek
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Hapus data Usulan Proyek (Soft Delete)
      tags:
      - Usulan Proyek
    get:
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Usulan Proyek
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Ambil detail Usulan Proyek by ID
      tags:
      - Usulan Proyek
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID Usulan Proyek
        in: path
        name: id
        required: true
        type: string
      - description: Data yang akan diedit
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/transaction.RequestUsulanProyek'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Base'
      summary: Update data Usulan Proyek
      tags:
      - Usulan Proyek
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"